requirements.txt
# use the pip install -r requirements.txt to install 


torch>=1.7.0
torchvision
tensorboard
lpips
numpy
Pillow
matplotlib
pyyaml
scikit-image
